📦 Add Action type handlers
- Organize main.ts by modularizing its actions
- Better Error Handling for messages
