+++ src/main.ts
@@ -1,25 +1,18 @@
 import { 
-  PresenceUpdate, startBot, 
-  sendDirectMessage, getUser, getGuild,
-  Guild,
+  startBot, getUser, getGuild,
 } from 'https://deno.land/x/discordeno@11.2.0/mod.ts';
-import { v4 } from 'https://deno.land/std@0.100.0/uuid/mod.ts';
 import { config } from 'https://deno.land/x/dotenv@v2.0.0/mod.ts';
 import { IEnvironment } from './Interfaces/index.ts';
-import {
-  PrecenseLogModel, UserModel,
-  GuildModel, GuildActivityModel,
-  BotTrackerModel,
-  initConnection,
+import { 
+  UserModel, initConnection,
 } from './Database/index.ts';
-import {
-  IPrecenseLog, IUser, StatusType,
-  IBotTracker, IGuild,
-} from './Interfaces/Database.ts';
-import { Model } from 'https://deno.land/x/denodb@v1.0.38/lib/model.ts';
-import { parseCommand } from './Commands/index.ts';
-import { statusEnumFromString } from './Helpers/utils.ts';
+import { IUser, IGuild } from './Interfaces/Database.ts';
 import { Cache } from './Helpers/Cache.ts';
+
+// Actions
+import { updateUserPresence } from './Actions/Presence.ts';
+import { checkAndNotifyBotTracking } from './Actions/Notifications.ts';
+import { handleGuildMessage } from './Actions/Messages.ts';
 
 // Logging System
 import Logger from './Logging/index.ts';
@@ -38,97 +31,6 @@ Log.Info('Database Connected!');
 export const GUILD_CACHE = new Cache<IGuild>(5);
 export const GUILD_CACHE_TTL = 24 * 60 * 60 * 1000; // 24 Hours
 
-
-async function updateUserPrecense(user: Model, precense: PresenceUpdate) {
-  // Check if there is a pending Status
-  const entryResult = await PrecenseLogModel
-    .where('userID', user.userID as string)
-    .orderBy('created_at', 'desc')
-    .limit(1)
-    .get();
-
-  // Setup Status ID
-  const status: StatusType = statusEnumFromString(precense.status);
-
-  // Close off Entry
-  if (entryResult.length) {
-    // Update ONLY if status type is Different
-    const pEntry: IPrecenseLog & Model = (entryResult as any)[0];
-    const pStatusID = parseInt(pEntry.statusID as any);
-
-    // Status Differs
-    if (pEntry.endTime === null && pStatusID !== status) {
-      PrecenseLogModel
-        .where('precenseID', pEntry.precenseID)
-        .update({ endTime: new Date().toUTCString() })
-        .then(() => Log.Info(`User ${user.userID} precense log updated.`))
-        .catch(err => Log.Error('User Precense\'s Endtime could not be updated.', err));
-    } 
-    
-    // Same as Entry, no new precense to log
-    else if (pEntry.endTime === null && pStatusID === status) {
-      Log.Print('No new precense to log. Same as before.');
-      return;
-    }
-  }
-
-  // Create a new Entry
-  const uuid = v4.generate().split('-').pop();
-  PrecenseLogModel.create({
-    precenseID: uuid,
-    userID: user.userID,
-    statusID: status,
-    startTime: new Date().toUTCString(),
-    endTime: null,
-  } as any)
-    .then(plog => Log.Info('Precense Log Created for ', user.userID))
-    .catch(err => Log.Error('Precense Log could not be created: ', err));
-}
-
-async function checkAndNotifyBotTracking(botUser: IUser, newPresence: string) {
-  Log.Internal('checkAndNotifyBotTracking', 'Notifying users of bot presence change');
-
-  // Get tracking entries
-  const tracking_entries: IBotTracker[] = await BotTrackerModel
-    .where('botId', botUser.userID)
-    .get() as any;
-  
-  // Check if Presence Changed
-  const status: StatusType = statusEnumFromString(newPresence);
-  const entryResult = await PrecenseLogModel
-    .where('userID', botUser.userID)
-    .orderBy('created_at', 'desc')
-    .limit(1)
-    .get();
-
-  if (entryResult.length) {
-    const pEntry: IPrecenseLog & Model = (entryResult as any)[0];
-    const pStatusID = parseInt(pEntry.statusID as any);
-
-    // Same as Entry, no new precense to log
-    if (pEntry.endTime === null && pStatusID === status) {
-      Log.Print('No new precense for bot. Not notifying anyone.');
-      return;
-    }
-  }
-  
-  // Presence Changed, notify all
-  Log.Info(`Notifying ${tracking_entries.length} users of bot ${botUser.username}[${botUser.userID}] presence change`);
-  for (const entry of tracking_entries) {
-    return sendDirectMessage(BigInt(entry.userId), `**${botUser.username}[${botUser.userID}]**: Presence Changed to \`${newPresence}\``);
-  }
-}
-
-async function addGuild(guild: Guild) {
-  // Store Guild Entry
-  return GuildModel.create({
-    guildID: guild.id,
-    guildName: guild.name,
-    responseChannel: null,
-  })
-    .then(() => Log.Info(`Guild Added: ${guild.name}`))
-    .catch(err => Log.Error('Guild Model Create Error: ', err));
-}
 
 // Initialize Bot
 let gatewayReady = false;
@@ -153,65 +55,11 @@ startBot({
         Log.Info(`Message did not originate from a Guild. GuildID = ${msg.guildId}`);
         return;
       }
-      
-      const { content, channel } = msg;
-      const author = await getUser(msg.authorId);     // TODO: Cache me!
 
-      // Check Guild in Cache
-      let cached_guild = GUILD_CACHE.get(msg.guildId.toString());
-
-      try {
-        if (!cached_guild) {
-          // Make sure Guild is stored in DB
-          const guild_entry = await GuildModel.where('guildID', msg.guildId.toString()).get();
-          if (!guild_entry.length) {
-            Log.Error('Guild ID not found: ', msg.guildId);
-            const guild = await getGuild(msg.guildId);
-            GUILD_CACHE.set(msg.guildId.toString(), guild as any, GUILD_CACHE_TTL);
-            await addGuild(guild as any);
-            cached_guild = guild as any;
-          } else {
-            GUILD_CACHE.set(msg.guildId.toString(), (guild_entry as any)[0], GUILD_CACHE_TTL);
-            cached_guild = (guild_entry as any)[0];
-          }
-        }
-      } catch(err) {
-        Log.Error('Message Create Guild Cache Error: ', err);
-      }
-
-      // Handle message only in verified channels
-      if (cached_guild?.responseChannel === null || channel?.name === cached_guild?.responseChannel) {
-        // Extract/Confirm Valid Command
-        if (content.startsWith('!')) {
-          Log.Info(`${author.username} issued command: ${content}`);
-
-          const command = parseCommand(content);
-          if (command) {
-            // Add Additional Command Information
-            command.userId = author.id;
-
-            command?.execute(msg, command)
-              .then(async () => {   // Store Executed Command from Server
-                const uuid = v4.generate().split('-').pop();
-                const combined_cmd = command.cmd + (command.arguments.length
-                  ? ' ' + command.arguments.join(' ') : '');
-
-                GuildActivityModel.create({
-                  guildActivityID: uuid as string,
-                  guildID: msg.guildId.toString(),
-                  command: combined_cmd,
-                })
-                  .then(() => Log.Info(`Guild Activity Added to ${msg.guildId}: ${combined_cmd} -> ${uuid}`))
-                  .catch(err => Log.Error('Guild Activity Error:', err));
-              })
-              .catch(err => {
-                Log.Error('Error:', err);
-                msg.reply(`🐞 Something bad happend! Please report to Devs. Timestamp: ${Date.now()}`)
-                  .catch(err => Log.Error('Message Reply Error:', err));
-                return;
-              });
-          }
-        }
+      // Guild Messages
+      else {
+        handleGuildMessage(msg)
+          .catch(err => Log.Error('Uncaught Exception <handleGuildMessage>: ', err));
       }
     },
 
@@ -232,16 +80,16 @@ startBot({
             // Fetch ALL Data for User
             const userPayload = await getUser(BigInt(precense.user.id));
             Log.Info(`Adding ${userPayload.username}[${userPayload.id}] to Database`);
-            
+
             // Add New Precense
             UserModel.create({
               userID: userPayload.id,
               username: userPayload.username,
               disableTracking: null,
               isBot: userPayload.bot,
-            } as Partial<IUser>)
+            } as any)
               // Update PrecenseLog
-              .then(user => updateUserPrecense(user, precense))
+              .then(user => updateUserPresence(user, precense))
               .catch(err => Log.Error('User Creation Error:', err));
           }
 
@@ -266,7 +114,7 @@ startBot({
             
             // Update PrecenseLog if User has an unclosed Precense & Did not disable tracking
             if ((user as any as IUser).disableTracking !== true) {
-              updateUserPrecense(user, precense);
+              updateUserPresence(user, precense);
             }
           }
